+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 8.30.0
 ------
 * disentangle order of states in `HabituationChoiceWorldSession`
+* fix ITI durations in `ChoiceWorldSession` and `HabituationChoiceWorldSession`
 
 8.29.0
 ------